requirements.txt
wxPython
requests
beautifulsoup4
pycrypto
ntplib
mechanize
httplib2
selenium
Pillow